Catch StorageNotAvailableException while formatting the share
The getById method called while formatting could go down to the storage
if somehow the node needs to be updated. If so, the storage where the
node is could not be available, causing the exception to pop up.
Such exception causes the list of shares to be broken since it wasn't
being caught.